Resolved 5 conflicts:

- site/data/noisy-{16384,4096,512}.json: removed (noisy profile was
  deleted on main — the PR's additions to these files are moot)
- site/data/current.json: picked main's commit hash (file is
  auto-regenerated by benchmark.sh --save anyway)
- site/data/frameworks.json: kept swerver entry (the PR's core
  addition); dropped orphan starlette/uwebsockets/vertx entries
  (no framework directories on either side); used main's workerman
  entry (newer, from PR #470)

Also removed "noisy" from frameworks/swerver/meta.json tests array
since that profile no longer exists.

Auto-merged the rest: absorbed main's contributor PRs (workerman,
rails/roda/ngx-php/fastpysgi/swoole updates, json-tls profile, actix
tuning, site data refreshes, etc.) while preserving the swerver
framework files and its baseline/pipelined/limited-conn/baseline-h2
benchmark result entries.

+++ frameworks/sisk/Model.cs
@@ -5,7 +5,7 @@ public class DatasetItem
     public int Id { get; set; }
     public string Name { get; set; } = "";
     public string Category { get; set; } = "";
-    public double Price { get; set; }
+    public int Price { get; set; }
     public int Quantity { get; set; }
     public bool Active { get; set; }
     public List<string>? Tags { get; set; }
@@ -17,17 +17,17 @@ public class ProcessedItem
     public int Id { get; set; }
     public string Name { get; set; } = "";
     public string Category { get; set; } = "";
-    public double Price { get; set; }
+    public int Price { get; set; }
     public int Quantity { get; set; }
     public bool Active { get; set; }
     public List<string>? Tags { get; set; }
     public RatingInfo? Rating { get; set; }
-    public double Total { get; set; }
+    public long Total { get; set; }
 }
 
 public class RatingInfo
 {
-    public double Score { get; set; }
+    public int Score { get; set; }
     public int Count { get; set; }
 }
