+++ frameworks/spring-boot/src/main/java/com/httparena/spring/boot/Rating.java
@@ -1,3 +1,3 @@
 package com.httparena.spring.boot;
 
-public record Rating(double score, long count) {}
+public record Rating(int score, long count) {}
